4.21*: Blog List Expansion, step 9

+++ part4/bloglist/controllers/blogs.js
@@ -61,6 +61,29 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!decodedToken) {
+    return response.status(401).json({error: 'token invalid'})
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (!blog) {
+    return response.status(404).json({error: 'blog not found'})
+  }
+
+  /*
+  if (blog.user.toString() !== decodedToken.id.toString()) {
+    return response.status(401).json({ 
+      error: 'only the creator can delete this blog' 
+    })
+  } */
+
+  if (!blog.user.equals(decodedToken.id)) {
+    return response.status(401).json({error: 'unauthorized'})
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
